requirements.txt
flask
openai
python-dotenv
spacy
PyPDF2
gunicorn
